requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
arrow==1.2.1; python_version >= '3.6'
asgiref==3.4.1; python_version >= '3.6'
backports.zoneinfo==0.2.1; python_version < '3.9'
blessed==1.19.0; python_version >= '2.7'
django-picklefield==3.0.1; python_version >= '3'
django-q==1.3.9
django==4.0
djangorestframework==3.13.1
gitdb==4.0.9; python_version >= '3.6'
gitpython==3.1.24
pyasn1==0.4.8
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2021.3
redis==3.5.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
rsa==4.8
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
smmap==5.0.0; python_version >= '3.6'
sqlparse==0.4.2; python_version >= '3.5'
typing-extensions==4.0.1; python_version < '3.10'
wcwidth==0.2.5
